[MODORG-42]-Replaced old HttpClient to WebClient.
