+++ src/main/java/org/folio/rest/client/RestClient.java
@@ -3,21 +3,12 @@ package org.folio.rest.client;
 import static javax.ws.rs.core.HttpHeaders.LOCATION;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
-import static org.apache.commons.lang3.StringUtils.EMPTY;
-import static org.apache.commons.lang3.StringUtils.isEmpty;
 import static org.folio.config.Constants.OKAPI_URL;
-import static org.folio.util.ResourcePathResolver.ACQUISITIONS_MEMBERSHIPS;
-import static org.folio.util.ResourcePathResolver.ACQUISITIONS_UNITS;
-import static org.folio.util.ResourcePathResolver.CQL_SORT_BY_PATTERN;
-import static org.folio.util.ResourcePathResolver.resourcesPath;
+import static org.folio.util.RestUtils.ID;
+import static org.folio.util.RestUtils.SUCCESS_RESPONSE_PREDICATE;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
-import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.CompletionException;
-import java.util.regex.Matcher;
 
 import io.vertx.core.Future;
 import io.vertx.core.MultiMap;
@@ -26,53 +17,14 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.client.WebClientOptions;
-import io.vertx.ext.web.client.predicate.ErrorConverter;
-import io.vertx.ext.web.client.predicate.ResponsePredicate;
-import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
-import org.folio.exception.HttpException;
 import org.folio.okapi.common.WebClientFactory;
 
 import io.vertx.core.Context;
 import io.vertx.core.json.JsonObject;
-import one.util.streamex.StreamEx;
 
 public class RestClient {
-  public static final String SEARCH_PARAMS = "?limit=%s&offset=%s%s&lang=%s";
-  private static final String ID = "id";
-  public static final String ACQUISITIONS_UNIT_ID = "acquisitionsUnitId";
-  public static final String IS_DELETED_PROP = "isDeleted";
-  public static final String ALL_UNITS_CQL = IS_DELETED_PROP + "=*";
-  public static final String ACTIVE_UNITS_CQL = IS_DELETED_PROP + "==false";
-  public static final String ACQUISITIONS_UNIT_IDS = "acqUnitIds";
-  public static final String NO_ACQ_UNIT_ASSIGNED_CQL = "cql.allRecords=1 not " + ACQUISITIONS_UNIT_IDS + " <> []";
-  public static final String GET_UNITS_BY_QUERY = resourcesPath(ACQUISITIONS_UNITS) + SEARCH_PARAMS;
-  public static final String GET_UNITS_MEMBERSHIPS_BY_QUERY = resourcesPath(ACQUISITIONS_MEMBERSHIPS) + SEARCH_PARAMS;
-
-  private static final ErrorConverter ERROR_CONVERTER = ErrorConverter.createFullBody(
-    result -> new HttpException(result.response().statusCode(), result.response().bodyAsString()));
-  private static final ResponsePredicate SUCCESS_RESPONSE_PREDICATE = ResponsePredicate.create(ResponsePredicate.SC_SUCCESS, ERROR_CONVERTER);
-
-  public static String buildQuery(String query, Logger logger) {
-    return isEmpty(query) ? EMPTY : "&query=" + encodeQuery(query, logger);
-  }
-
-  /**
-   * @param query  string representing CQL query
-   * @param logger {@link Logger} to log error if any
-   * @return URL encoded string
-   */
-  public static String encodeQuery(String query, Logger logger) {
-    try {
-      return URLEncoder.encode(query, StandardCharsets.UTF_8.toString());
-    } catch (UnsupportedEncodingException e) {
-      logger.error("Error happened while attempting to encode '{}'", query, e);
-      throw new CompletionException(e);
-    }
-  }
-
-  /**
+   /**
    * A common method to create a new entry in the storage based on the Json Object.
    *
    * @param recordData json to post
@@ -210,39 +162,6 @@ public class RestClient {
       id = location.substring(location.lastIndexOf('/') + 1);
     }
     return id;
-  }
-
-  public static String combineCqlExpressions(String operator, String... expressions) {
-    if (ArrayUtils.isEmpty(expressions)) {
-      return EMPTY;
-    }
-
-    String sorting = EMPTY;
-
-    // Check whether last expression contains sorting query. If it does, extract it to be added in the end of the resulting query
-    Matcher matcher = CQL_SORT_BY_PATTERN.matcher(expressions[expressions.length - 1]);
-    if (matcher.find()) {
-      expressions[expressions.length - 1] = matcher.group(1);
-      sorting = matcher.group(2);
-    }
-
-    return StreamEx.of(expressions)
-      .filter(StringUtils::isNotBlank)
-      .joining(") " + operator + " (", "(", ")") + sorting;
-  }
-
-  /**
-   * Transform list of id's to CQL query using 'or' operation
-   * @param ids list of id's
-   * @return String representing CQL query to get records by id's
-   */
-  public static String convertIdsToCqlQuery(Collection<String> ids) {
-    return convertIdsToCqlQuery(ids, ID, true);
-  }
-
-  public static String convertIdsToCqlQuery(Collection<String> values, String fieldName, boolean strictMatch) {
-    String prefix = fieldName + (strictMatch ? "==(" : "=(");
-    return StreamEx.of(values).joining(" or ", prefix, ")");
   }
 
   private MultiMap convertToCaseInsensitiveMap(Map<String, String> okapiHeaders) {
