+++ src/test/java/org/folio/rest/impl/MockServer.java
@@ -21,21 +21,21 @@ import static org.folio.rest.impl.TestEntities.ORGANIZATION_FULL_PROTECTED;
 import static org.folio.rest.impl.TestEntities.ORGANIZATION_NO_ACQ;
 import static org.folio.rest.impl.TestEntities.ORGANIZATION_READ_PROTECTED;
 import static org.folio.rest.impl.TestEntities.ORGANIZATION_UPDATE_PROTECTED;
-import static org.folio.rest.client.RestClient.ACQUISITIONS_UNIT_ID;
-import static org.folio.rest.client.RestClient.ACQUISITIONS_UNIT_IDS;
-import static org.folio.rest.client.RestClient.ACTIVE_UNITS_CQL;
-import static org.folio.rest.client.RestClient.ALL_UNITS_CQL;
-import static org.folio.rest.client.RestClient.GET_UNITS_MEMBERSHIPS_BY_QUERY;
-import static org.folio.rest.client.RestClient.NO_ACQ_UNIT_ASSIGNED_CQL;
-import static org.folio.rest.client.RestClient.SEARCH_PARAMS;
-import static org.folio.rest.client.RestClient.buildQuery;
-import static org.folio.rest.client.RestClient.combineCqlExpressions;
-import static org.folio.rest.client.RestClient.convertIdsToCqlQuery;
 import static org.folio.service.organization.OrganizationStorageService.GET_ORGANIZATIONS_BY_QUERY;
 import static org.folio.util.ResourcePathResolver.ACQUISITIONS_UNITS;
 import static org.folio.util.ResourcePathResolver.ORGANIZATIONS;
 import static org.folio.util.ResourcePathResolver.resourceByIdPath;
 import static org.folio.util.ResourcePathResolver.resourcesPath;
+import static org.folio.util.RestUtils.ACQUISITIONS_UNIT_ID;
+import static org.folio.util.RestUtils.ACQUISITIONS_UNIT_IDS;
+import static org.folio.util.RestUtils.ACTIVE_UNITS_CQL;
+import static org.folio.util.RestUtils.ALL_UNITS_CQL;
+import static org.folio.util.RestUtils.GET_UNITS_MEMBERSHIPS_BY_QUERY;
+import static org.folio.util.RestUtils.NO_ACQ_UNIT_ASSIGNED_CQL;
+import static org.folio.util.RestUtils.SEARCH_PARAMS;
+import static org.folio.util.RestUtils.buildQuery;
+import static org.folio.util.RestUtils.combineCqlExpressions;
+import static org.folio.util.RestUtils.convertIdsToCqlQuery;
 
 import java.util.Arrays;
 import java.util.Objects;
